fix: post-u128-cascade cleanup — cfg/reader test paths, ir test layout, opt bench
After the IntConst u128 cascade landed, several auxiliary test files and
the file structure of crates/ir/src/node/output_type.rs needed follow-up:

  cfg/tests/common/real_binary.rs:
    - binary() now takes (arch, fn_name) and resolves the per-category
      fixture (.elf) via a function-to-category map; the legacy single
      test.elf was deleted by the analyzer-crate review.

  cfg/tests/cfg_integration.rs:
    - All cfg_of()/binary() calls pass the function name.
    - Function references updated: bitwise_ops -> bit_and, fib -> fib_recursive,
      g dropped (no longer in fixtures).

  cfg/tests/{dot_dumper,cfg_query,vn_to_name}.rs:
    - binary("x64") -> binary("x64", fn_name) (or binary("x64", "add")
      where fn_name was hardcoded).

  reader/tests/elf_smoke.rs:
    - Smoke tests use arithmetic.elf (was the deleted test.elf).

  ir/src/node/output_type.rs:
    - Move the #[cfg(test)] mod tests block to the end of the file so
      no items appear after it (clippy items_after_test_module).

  opt/src/constant_fold/rules.rs:
    - lzcount fold rule guards against U256 inputs (returns skip);
      previously panicked with "attempt to subtract with overflow"
      on (128 - 256).  Test fold_lzcount_u256_input_skips_cleanly
      now passes.

  opt/benches/stack_store.rs:
    - bare loop variable `i` (already u64) doesn't need u64::from().

Final workspace state: 1242 passed, 0 failed, 137 ignored, clippy clean.

+++ crates/opt/benches/stack_store.rs
@@ -32,7 +32,7 @@ fn build_pushes(n: usize) -> ir::BuiltFunctionGraph {
             .build_int_binary_operation(sp_v, four, IntBinaryOp::Sub, NodeOutputType::U32)
             .unwrap();
         b.write_variable(&sp, sp_v).unwrap();
-        let data = b.build_int_const(u64::from(i), NodeOutputType::U32);
+        let data = b.build_int_const(i, NodeOutputType::U32);
         b.build_store(sp_v, data, rsleigh::VnSpace::RAM).unwrap();
     }
     b.build_return(None, &[]).unwrap();
